requirements.txt
anthropic>=0.18.0
pytesseract>=0.3.10
Pillow>=10.0.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0
python-dateutil>=2.8.2
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
